Refactor temporary model creation in specs.

+++ spec/helpers/temporary_models.rb
@@ -4,43 +4,73 @@ module TemporaryModels
   end
 
   #
-  # Create a table and model for the duration of the given block.
+  # Set up a model class with the given name for all examples in this
+  # example group.  +columns+ is a hash from column name to
+  # ActiveRecord type (e.g., :string or :integer).
   #
-  # +columns+ is a map from column name to type.
+  # The class +name+ may be a String or Symbol.  It can also be a
+  # singleton hash mapping the class name to the superclass name.  If
+  # the superclass is not ActiveRecord::Base, it is assumed to
+  # indicate single table inheritance, and no table will be created.
   #
-  def with_model(model_name, columns={})
-    table_name = model_name.to_s.underscore.pluralize
-    ActiveRecord::Base.connection.create_table(table_name) do |table|
-      columns.each do |column_name, column_type|
-        table.send column_type, column_name
+  def create_model_class(name, columns={}, &class_body)
+    if name.is_a?(Hash) && name.size == 1
+      class_name, superclass_name = *name.to_a.flatten
+    else
+      class_name, superclass_name = name, 'ActiveRecord::Base'
+    end
+    need_table = superclass_name.to_s == 'ActiveRecord::Base'
+
+    table_name = class_name.to_s.underscore.pluralize
+    if need_table
+      ActiveRecord::Base.connection.create_table(table_name) do |table|
+        columns.each do |column_name, column_type|
+          table.send column_type, column_name
+        end
       end
     end
-    Object.const_set(model_name, Class.new(ActiveRecord::Base))
+    klass = Class.new(superclass_name.to_s.constantize, &class_body)
+    Object.const_set(class_name, klass)
+  end
+
+  #
+  # Destroy a model class created with #create_model_class, and drop
+  # the created table.  +name+ should be the same as the first
+  # argument given on creation.
+  #
+  def destroy_model_class(name)
+    if name.is_a?(Hash) && name.size == 1
+      class_name, superclass_name = *name.to_a.flatten
+    else
+      class_name, superclass_name = name, 'ActiveRecord::Base'
+    end
+    need_table = superclass_name.to_s == 'ActiveRecord::Base'
+
+    table_name = class_name.to_s.underscore.pluralize
+    ActiveRecord::Base.connection.drop_table(table_name) if need_table
+    Object.send(:remove_const, class_name)
+  end
+
+  #
+  # Create a model and table for the duration of the block.  See
+  # #create_model_class for the meaning of the arguments.
+  #
+  def with_model_class(name, columns={})
+    create_model_class(name, columns)
     yield
   ensure
-    ActiveRecord::Base.connection.drop_table(table_name)
-    Object.send(:remove_const, model_name)
+    destroy_model_class(name)
   end
 
   module ClassMethods
     #
-    # Set up a model class with the given name for all examples in this
-    # example group.  You may pass a block to configure the database
-    # table like an ActiveRecord migration.
+    # Create a model and table for the duration of each example in
+    # this example group.  See #create_model_class for the meaning of
+    # the arguments.
     #
-    def set_up_model_class(name, superclass_name='ActiveRecord::Base', &block)
-      need_table = superclass_name == 'ActiveRecord::Base'
-      block ||= lambda{}
-
-      before do
-        ActiveRecord::Base.connection.create_table(name.to_s.underscore.pluralize, &block) if need_table
-        Object.const_set(name, Class.new(superclass_name.to_s.constantize))
-      end
-
-      after do
-        Object.send(:remove_const, name)
-        ActiveRecord::Base.connection.drop_table(name.to_s.underscore.pluralize) if need_table
-      end
+    def use_model_class(name, columns={}, &class_body)
+      before{create_model_class(name, columns, &class_body)}
+      after{destroy_model_class(name)}
     end
   end
 end
